requirements.txt
cssselect==1.3.0
greenlet==3.2.3
jmespath==1.0.1
lxml==6.0.0
nested-lookup==0.2.25
packaging==25.0
parsel==1.10.0
playwright==1.54.0
psutil==7.0.0
pyee==13.0.0
python-dotenv==1.1.1
six==1.17.0
typing_extensions==4.14.1
w3lib==2.3.1
